futures: only run tests depending on feature flags
## Motivation

Previously, the tests for `std::future` support were always compiled,
regardless of whether or not the `std::future` feature was enabled. This
is because the tests were in the integration tests dir, and were not
feature flagged.

## Solution

I've moved these tests into modules in the lib tests, so that they can
only be compiled when the appropriate feature flags are enabled. I've
also made some tweaks to the test support code to accomodate this.
