+++ tracing/tests/support/metadata.rs
@@ -9,7 +9,7 @@ pub struct Expect {
 }
 
 impl Expect {
-    pub(in crate::support) fn check(&self, actual: &Metadata<'_>, ctx: fmt::Arguments<'_>) {
+    pub(super) fn check(&self, actual: &Metadata<'_>, ctx: fmt::Arguments<'_>) {
         if let Some(ref expected_name) = self.name {
             let name = actual.name();
             assert!(
