+++ tracing/tests/support/span.rs
@@ -8,14 +8,14 @@ use std::fmt;
 /// `subscriber` module.
 #[derive(Clone, Debug, Default, Eq, PartialEq)]
 pub struct MockSpan {
-    pub(in crate::support) metadata: metadata::Expect,
+    pub(super) metadata: metadata::Expect,
 }
 
 #[derive(Debug, Default, Eq, PartialEq)]
 pub struct NewSpan {
-    pub(in crate::support) span: MockSpan,
-    pub(in crate::support) fields: field::Expect,
-    pub(in crate::support) parent: Option<Parent>,
+    pub(super) span: MockSpan,
+    pub(super) fields: field::Expect,
+    pub(super) parent: Option<Parent>,
 }
 
 pub fn mock() -> MockSpan {
@@ -97,7 +97,7 @@ impl MockSpan {
         }
     }
 
-    pub(in crate::support) fn check_metadata(&self, actual: &tracing::Metadata<'_>) {
+    pub(super) fn check_metadata(&self, actual: &tracing::Metadata<'_>) {
         self.metadata.check(actual, format_args!("span {}", self));
         assert!(actual.is_span(), "expected a span but got {:?}", actual);
     }
